refactor(agent): drop dead code left by trajectory-manager refactor
Remove vestigial bookkeeping the turn-node TrajectoryTree made redundant:

* anthropic adapter: the always-empty dispatch_id plumbing in
  _anthropic_blocks / _build_reply (routing is now done by the tree, not
  by tool_use ids).
* hoist the byte-identical Session dataclass and finish_session method
  from both adapters into common.BaseAdapter (shared session_cls +
  export_token_segments drain).
* trajectory_manager: delete the unreferenced _starting_chains /
  _leaf_of_chain helpers.

No behavior change; agent adapter and trajectory tests pass.
